Various parsing and logic fixes Fix missing recvPciProtocol in SCardTransmitRequest::packPayload Fix incorrect isEof condition in ContentCopyControlDescriptor::unpack Fix incorrect type in Ddmt::unpack Fix incorrect parsing of eventMsgGroupId in Emt::unpack Fix operator precedence bugs in MhAudioComponentDescriptor Fix incorrect return value in MhCacheControlInfoDescriptor::unpack Fix build error by adding missing cstddef include in pesPacket.h Fix incorrect bit operations in SignalingMessage::unpack Standardize error code output for smart card errors Fix incorrect return value in ReadStream::peek Remove unused variables

+++ src/ddmt.cpp
@@ -63,7 +63,7 @@ bool Ddmt::unpack(Common::ReadStream& stream) {
         baseDirectoryPath.resize(baseDirectoryPathLength); 
         stream.read(baseDirectoryPath.data(), baseDirectoryPathLength);
 
-        uint16_t numOfDirectoryNodes = stream.get8U();
+        uint8_t numOfDirectoryNodes = stream.get8U();
         for (int i = 0; i < numOfDirectoryNodes; i++) {
             Node node;
             if (!node.unpack(stream)) {
